almost at hair tearing situation here. no idea whats going on under the hood with vitest but one moment its fine and seeing esm then middle of the second set of tests it loses ability to see esm on certain files, then regains it but loses another. spent 3+ hours going through this in the most painful detail to get that running again but the AI tooling just of note, really struggled to comprehend issues and sent me in circles trying to resolve that . got close now but had to revert work done yesterday where type module was added to align all scripts now i need to specify some files as mjs because esbuild will just NOT see them as esm to compile and run. To make this any easier or better in future I would think it better to stick to ts functions then compile it all to a dist folder but that wasnt tuned in from the beginning and too late to add now.  also the AI tooling while i think of it hallucinated or just absolutely incorrectly did not capture a double export as well, no idea how that flew under the radar until I caught it.
other overall changes, split some of the code components out into punctional and logic so that vitest can reach them better. I had a working cover for all 3 but two of them didn't report in as they were in a subprocess so that was annoying as coverage was 0 so developed a second cover which spawned all these test issues. now to go back and retro fit the original commands so they work as expected / described